network: configure the network interface to block until it has an address
This should prevent issues where some services (like cloud-init) start
because "networking.service" completed, although the main network interface
is still waiting for an address.

There could be a race condition at that time: if the interface gets an
address really quickly, cloud-init may see it already configured, but most
of the time it's not.

This is more similar to what the default Debian AWS EC2 image does as well.